+++ src/game/nlcore/action/actions/persistentAction.ts
@@ -2,7 +2,9 @@ import {PersistentActionContentType, PersistentActionTypes} from "@core/action/a
 import {GameState} from "@player/gameState";
 import {TypedAction} from "@core/action/actions";
 import {Values} from "@lib/util/data";
-import {Persistent} from "@core/elements/persistent";
+import {Persistent, PersistentContent} from "@core/elements/persistent";
+import { Namespace } from "../../common/game";
+import { ContentNode } from "../tree/actionTree";
 
 export class PersistentAction<T extends Values<typeof PersistentActionTypes> = Values<typeof PersistentActionTypes>>
     extends TypedAction<PersistentActionContentType, T, Persistent<any>> {
@@ -11,7 +13,7 @@ export class PersistentAction<T extends Values<typeof PersistentActionTypes> = V
     executeAction(gameState: GameState) {
         const action: PersistentAction = this;
         if (action.is<PersistentAction<"persistent:set">>(PersistentAction, "persistent:set")) {
-            const [key, value] = action.contentNode.getContent();
+            const [key, value] = (action.contentNode as ContentNode<PersistentActionContentType["persistent:set"]>).getContent();
             const namespace = gameState.getStorable().getNamespace(
                 action.callee.getNamespaceName()
             );
@@ -29,8 +31,26 @@ export class PersistentAction<T extends Values<typeof PersistentActionTypes> = V
             }, [prevValue]);
 
             return super.executeAction(gameState);
-        }
+        } else if (action.is<PersistentAction<"persistent:assign">>(PersistentAction, "persistent:assign")) {
+            const [value] = (action.contentNode as ContentNode<PersistentActionContentType["persistent:assign"]>).getContent() as [Partial<PersistentContent>];
+            const namespace = gameState.getStorable().getNamespace(
+                action.callee.getNamespaceName()
+            ) as Namespace<PersistentContent>;
+            const prevValue: Partial<PersistentContent> = {};
 
+            Object.keys(value).forEach((key: string) => {
+                prevValue[key] = namespace.get(key);
+                namespace.set(key, value[key]);
+            });
+
+            gameState.actionHistory.push<[any]>(this, (prevValue) => {
+                Object.keys(prevValue).forEach(key => {
+                    namespace.set(key, prevValue[key]);
+                });
+            }, [prevValue]);
+            
+            return super.executeAction(gameState);
+        }
         throw this.unknownTypeError();
     }
 }